+++ packages/airgap.js-types/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @transcend-io/airgap.js-types
 
+## 14.2.7
+
+### Patch Changes
+
+- Updated dependencies [9b1c5f3]
+  - @transcend-io/internationalization@4.1.1
+  - @transcend-io/privacy-types@5.2.3
+
+## 14.2.6
+
+### Patch Changes
+
+- Updated dependencies [c14ba60]
+  - @transcend-io/privacy-types@5.2.2
+
 ## 14.2.5
 
 ### Patch Changes